Improve support for multiple steam library paths and fix some shortcut naming

+++ scripts/es_steam_import.sh
@@ -5,7 +5,7 @@
 # File Created: Monday, 11th September 2023 3:57:47 pm
 # Author: Josh.5 (jsunnex@gmail.com)
 # -----
-# Last Modified: Saturday, 16th September 2023 10:33:10 am
+# Last Modified: Saturday, 16th September 2023 6:52:35 pm
 # Modified By: Josh.5 (jsunnex@gmail.com)
 ###
 #
@@ -20,17 +20,15 @@
 
 set -euo pipefail
 
-
 # Import helpers
 source "${USER_HOME:?}/init.d/helpers/functions.sh"
-
 
 # Ensure this script is being executed as the default user
 exec_script_as_default_user
 
 function CreateXDGDesktopShorcut() {
     __steam_id="$1"
-    __steam_name="$2"
+    __steam_name="${@:2}"
 
     cat <<EOF
 [Desktop Entry]
@@ -57,20 +55,36 @@ for steamPath in ${steamPaths}; do
         echo "Ignoring steam library '${steamPath:?}'"
         continue
     fi
+    if [ ! -d "${steamPath:?}/steamapps" ]; then
+        echo "Ignoring steam library '${steamPath:?}' as it does not have any steamapps directory."
+        continue
+    fi
+    if ! pushd "${steamPath:?}/steamapps/" &>/dev/null; then
+        echo "Failed to change directory to '${steamPath:?}/steamapps/'. Ignoring path."
+        continue
+    fi
     echo "Parsing steam library '${steamPath:?}'"
-    __steamapps=$(ls "${steamPath:?}/steamapps/" | grep ".acf")
 
-    for __steamapp in ${__steamapps:?}; do
-        steam_id=$(grep "appid" "${steamPath:?}/steamapps/${__steamapp}" | cut -d '"' -f 4)
-        steam_name_dirty=$(grep "name" "${steamPath:?}/steamapps/${__steamapp}" | cut -d '"' -f 4 | sed  -e 's/"//g' -e 's/^[[:space:]]*//' -e 's/[[:space:]]*$//')
-        steam_name_clean=$(echo "${steam_name_dirty:?}" | sed -e 's/[^a-zA-Z0-9_.-]/_/g' -e 's/™//g')
-        print_step_header "Found steam app '${steam_name_dirty:?}'"
-        if [[ ! -z "${steam_name_dirty:?}" ]]; then
-            print_step_header "Creating shortcut for steam app '${steam_name_dirty:?}'"
-            es_entry="${romsPath:?}/${steam_name_clean:?}.desktop"
-            es_shortcut=$(CreateXDGDesktopShorcut ${steam_id:?} ${steam_name_dirty:?})
+    # List .acf files in the directory and store them in an array
+    __acf_files=(*.acf)
 
-            echo "${es_shortcut:?}" > "${es_entry:?}"
+    # Loop over each .acf file and read its contents
+    for acf_file in "${__acf_files[@]}"; do
+        if [ -f "${acf_file:?}" ]; then
+            steam_id=$(grep '"appid"' "${acf_file:?}" | cut -d '"' -f 4)
+            steam_name_dirty=$(grep "name" "${acf_file:?}" | cut -d '"' -f 4 | sed -e 's/^[[:space:]]*//' -e 's/[[:space:]]*$//')
+            steam_name_clean=$(echo "${steam_name_dirty:?}" | sed -e 's/[^a-zA-Z0-9_.-:]/_/g' -e 's/™//g')
+
+            print_step_header "Found steam app '${steam_name_dirty:?}'"
+            if [[ ! -z "${steam_name_dirty:?}" ]]; then
+                print_step_header "Creating shortcut for steam app '${steam_name_dirty:?}'"
+                es_entry="${romsPath:?}/${steam_name_clean:?}.desktop"
+                es_shortcut=$(CreateXDGDesktopShorcut ${steam_id:?} ${steam_name_dirty:?})
+
+                echo "${es_shortcut:?}" >"${es_entry:?}"
+            fi
         fi
     done
+
+    popd &>/dev/null
 done
